+++ src/bin/cargo-pros.rs
@@ -1,6 +1,6 @@
 use cargo_pros::{build, launch_simulator, strip_binary};
 use clap::{Args, Parser, Subcommand};
-use std::path::PathBuf;
+use std::{path::PathBuf, process::Command};
 
 cargo_subcommand_metadata::description!("Manage pros-rs projects");
 
@@ -29,12 +29,41 @@ enum Commands {
         #[clap(last = true)]
         args: Vec<String>,
     },
+    Upload {
+        #[clap(long, short)]
+        slot: u8,
+        #[clap(long, short)]
+        file: Option<PathBuf>,
+        #[clap(long, short)]
+        action: UploadAction,
+
+        #[clap(last = true)]
+        args: Vec<String>,
+    },
     Sim {
         #[clap(long)]
         ui: Option<String>,
         #[clap(last = true)]
         args: Vec<String>,
     },
+}
+
+#[derive(Clone, Debug)]
+enum UploadAction {
+    Screen,
+    Run,
+    None,
+}
+impl std::str::FromStr for UploadAction {
+    type Err = String;
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "screen" => Ok(UploadAction::Screen),
+            "run" => Ok(UploadAction::Run),
+            "none" => Ok(UploadAction::None),
+            _ => Err("Invalid upload action".into()),
+        }
+    }
 }
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
@@ -48,6 +77,48 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
                     strip_binary(path);
                 }
             });
+        }
+        Commands::Upload {
+            slot,
+            file,
+            action,
+            args,
+        } => {
+            let mut artifact = None;
+            if let Some(path) = file {
+                artifact = Some(path);
+            } else {
+                let mut completed = false;
+                build(path.clone(), args, false, |new_artifact| {
+                    let mut bin_path = new_artifact.clone();
+                    bin_path.set_extension("bin");
+                    artifact = Some(bin_path.into());
+                    strip_binary(new_artifact);
+                    completed = true;
+                });
+                while !completed {
+                    std::thread::sleep(std::time::Duration::from_millis(10));
+                }
+            }
+            let artifact =
+                artifact.expect("Binary not found! Try explicitly providing one with --path (-p)");
+            Command::new("pros")
+                .args([
+                    "upload",
+                    "--target",
+                    "v5",
+                    "--slot",
+                    &slot.to_string(),
+                    "--after",
+                    match action {
+                        UploadAction::Screen => "screen",
+                        UploadAction::Run => "run",
+                        UploadAction::None => "none",
+                    },
+                    &artifact.to_string_lossy(),
+                ])
+                .spawn()?
+                .wait()?;
         }
         Commands::Sim { ui, args } => {
             let mut artifact = None;